formas-geometricas/requirements.txt
numpy==1.24.4
pygame==2.6.1
pyopengl==3.1.10